+++ src/components/Button/Button.js
@@ -2,8 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 
-import { flavorable } from '../../decorators';
-
 export const Button = (props) => {
   const {
     children,
@@ -14,12 +12,30 @@ export const Button = (props) => {
     title,
     tooltip,
     value,
+    neutral,
+    brand,
+    destructive,
+    inverse,
+    success,
+    iconSize,
+    iconBorder,
+    icon,
+    iconInverse,
     ...rest
   } = props;
 
   const sldsClasses = [
     'slds-button',
     { 'slds-is-selected': !!selected },
+    { 'slds-button_neutral': neutral },
+    { 'slds-button_brand': brand },
+    { 'slds-button_destructive': destructive },
+    { 'slds-button_inverse': inverse },
+    { 'slds-button_success': success },
+    { [`slds-button_icon-${iconSize}`]: !!iconSize },
+    { [`slds-button_icon-${iconBorder}`]: !!iconBorder },
+    { 'slds-button_icon': icon },
+    { 'slds-button_icon-inverse': iconInverse },
     className,
   ];
 
@@ -39,21 +55,6 @@ export const Button = (props) => {
   );
 };
 
-Button.flavors = [
-  'neutral',
-  'brand',
-  'destructive',
-  'icon',
-  'icon-border-filled',
-  'icon-container',
-  'icon-inverse',
-  'icon-border',
-  'icon-bare',
-  'icon-x-small',
-  'reset',
-  'success',
-];
-
 Button.defaultProps = {
   children: null,
   className: null,
@@ -62,6 +63,15 @@ Button.defaultProps = {
   title: null,
   tooltip: null,
   value: null,
+  neutral: false,
+  brand: false,
+  destructive: false,
+  inverse: false,
+  success: false,
+  icon: false,
+  iconSize: null,
+  iconBorder: null,
+  iconInverse: false,
 };
 
 Button.propTypes = {
@@ -97,6 +107,43 @@ Button.propTypes = {
    * adds optional value tag to the button
    */
   value: PropTypes.string,
+  /**
+   * neutral button
+   */
+  neutral: PropTypes.bool,
+  /**
+   * brand button
+   */
+  brand: PropTypes.bool,
+  /**
+   * destructive button
+   */
+  destructive: PropTypes.bool,
+  /**
+   * inverse button
+   */
+  inverse: PropTypes.bool,
+  /**
+   * success button
+   */
+  success: PropTypes.bool,
+  /**
+   * button has icon
+   */
+  icon: PropTypes.bool,
+  /**
+   * specify icon size if not medium
+   */
+  iconSize: PropTypes.oneOf(['xx-small', 'x-small', 'small', 'large']),
+  /**
+   * specify icon button border if any
+   */
+  iconBorder: PropTypes.oneOf(['border', 'border-filled', 'border-inverse']),
+  /**
+   * invert icon
+   */
+  iconInverse: PropTypes.bool,
+
 };
 
-export default flavorable(Button, 'button');
+export default Button;